PIRATES (#2142)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Rakshari get the pirate accent by default now
Also they can be swashbucklers
Also also they can be sand pirates
Unrelated but knocking on doors has a cooldown again, didn't wanna make
a whole new pr for that

![grafik](https://github.com/user-attachments/assets/4809f615-c70d-4405-b978-88e83c6265f3)

![grafik](https://github.com/user-attachments/assets/dcb7f293-4f01-439e-8f00-ce4a349bacb4)

![grafik](https://github.com/user-attachments/assets/8510e96f-967a-4d04-b0b8-971e2f55a087)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Accent was requested by Ook, swashbuckler seemed fine to me, desert
rider adds more variety/flavour to rakshari, otherwise they only get the
Zybantine Merc
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>