re-enable ping in stat panel and chat (#2247)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Enables server_maint to update client ping and re-enables the chatoutput
ehjax callback for ping.

Also re-enables clear chat option.


![image](https://github.com/user-attachments/assets/99f1abeb-26cb-44f6-aae6-a37c8a24ff43)


![image](https://github.com/user-attachments/assets/f69fdeee-5c40-4f12-888c-fa469c2e3fc2)


![image](https://github.com/user-attachments/assets/1e9388ca-ef55-4772-baa7-4a44d63c975d)


![image](https://github.com/user-attachments/assets/e58dcb29-e036-42c7-8c89-c11eea8739db)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
I let the players see the ping but it was never updated so now like 2
months later the fix is here woo

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->