Change some admin css (#2260)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
You can actually see the message for de/re-adminning

The watchlist notices aren't horrible contrast

JS fiddled without the ADMIN LOG:

before

![image](https://github.com/user-attachments/assets/c9c1cdd6-b760-4d1b-b14f-546fe1980760)

Now

![image](https://github.com/user-attachments/assets/bafc879e-1769-4fd4-ad7b-18be6b2f9f3b)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
It hurts my eyes

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->