Fixes and changes: Artificer recipe book, paintings management, witch hat, gods themes, mana overload, remove decree, family, VL mana regen (#2262)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

This PR does following fixes and changes:

- Forest guardian bait now uses its originally intended sprite instead
of being invisible
- Artificer recipe book now also contains all recipes you can make on an
artificer table, like prosthetics
- Bronze headhook can now be actually crafted on an artificer table, as
previously it did not have a set category-
- Fixes for player made paintings management UI, which did not even show
properly on live and runtimed locally
- Mailboxes without an explicitly set directory will now show their area
name on examine, to be consistent with a directory list
- Witch hat is now craftable, previously only wizard hat was craftable,
even if they differ only by sprite
- Added god omen sounds for Graggar, Baotha, Matthios by **Vispectra**
and Xylix by **DinoNugs** and made related god events use them. Also
reorganized current god related sounds to one sound folder
- Cleans up and tweaks personal events a little bit and also fixes a few
minor issues with them
- Fixes preferences UI size not being large enough to contain everything
without a scrollbar, and also fixes a few alignment issues with the
headers
- Hides Debug Loot Table verb from non admins
- Fixes OpenDream warnings
- Removes an unused variable in human helpers
- Fixes an issue where mana softcap increased way too high in some
cases, so any mana regeneration overloaded the mage eventually
- Added an ability to remove decrees the same way you can remove laws
from the throne, instead of them being permanent
- Added a cancel option to mana star item input menu, so you don't have
to write 0 just to close it
- Added a close button to manuscripts so you don't have to ALT + F4 to
close them
- Fixes pottery lathe examine and runtime
- Fixes several issues with the royal family, where members were set
incorrectly, so princess saw her father as their husband and so on
- Added a very small mana regen to a Vampire Lord, so he can use his
spells more than a few times in a round, requested by Huk
- Allows multiple of the same item to be used as a conduit for spells.
Previously, if the priest had two or more astrata amulets equipped,
spells would fail with a message that he lacked something, as the check
was for exactly one amulet, not more.

Fixes #1084
Fixes #2109
Fixes #2252
Fixes #2216
Fixes #1981
Fixes #1665
Fixes #2145
Most likely fixes #1918 (amulet requirement change)

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

Bugfixes and more complete content.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->