Schism and influence changes (#2261)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

This PR does a few changes to schism event and influences in general:

**Schism**

- Lowers event weight very considerably, should happen much less often
now
- Challenger is now decided by influence, not number of followers
- Challenger must have atleast 1 clergy of their faith present
- No triumph rewards for those who a choose side after the half point
announcement
- Loser deity also loses a bit of their influence
- If Astrata wins, her opponents will also get cursed by her to prevent
them getting healed by a miracles, priest should be able to lift these
curses individually if he decides to do so
- If Challenger wins, Astrata will throw a tantrum and hides the Sun for
a while
- Vice Priest can now only be chosen from the clergy and also gets a
passive devotion gain
- Eora will now not trigger the schism 

**Influences**
- Penalty from a repeated storyteller reigns is now slowly growing from
1 to a maximum of 2,5
- Adds a small follower penalty to Noc and Zizo, as they are often
forced patrons
- Minor influences balancing

Also updates census with a newly added tritons and cleans it up a
little.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

More meaningful schism and better influence balancing.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->