God Weapons Part I : The Inhumen. (#2308)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

This PR aims to add unique powerful weapons to the game. This is the
first part of many to follow including the Ten, Armors, and Areas for
crafting these.

The Pleonexia : 
Sword of legend. Fabled to steal _distance_... a cunning Matthiosite
always has a way out.
  
Gorefeast :
A hideous or beautiful axe made of bone and flesh. Strike fear into
their _hearts_, and drink deeply.

Neant : 
Neant or Void in imperial is a devastating weapon made for those with
the intelligence to make the most of it. They will draw their last
breath and steal their _soul_ away, for running will make no difference.

The Turbulenta : 
A bow, that also works as an instrument! _Spice_ up the battlefield AND
the court!

At the moment they are very very strong. Mostly for events and admins
until part II and III come out soon!
## Why It's Good For The Game

Every God should have their chosen weapon or item. It's flavorful and
helps make the gods in game feel more mystical.
Be warned however, ye faithful... they do not take lightly to those who
touch their relics... especially Zizo.

I don't want to spoil too much! They are supposed to be legendary after
all.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

![image](https://github.com/user-attachments/assets/c039cc54-838a-4ef9-a9e8-90dbf0a9d0ab)

![image](https://github.com/user-attachments/assets/318fd234-c47a-411a-8a84-b7aafcfde31f)

![image](https://github.com/user-attachments/assets/36768cab-f32b-4087-82e0-ed621b4fb8e2)

![image](https://github.com/user-attachments/assets/46e75d5f-2df0-44ce-acbe-db7cf94fd341)

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>