Hollow Ranger: A New Adventurer from Amber Hollow! Ft. Hollowkin, Humen, Harpies! (#2325)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
This class is an experiment in the idea of what a Ranger would be if
they were more focused on scouting for a would-be party of bandits than
on nature itself. Of course, for one reason or another, they've
abandoned the life, and now they've shown up in Vanderlin! The Hollow
Ranger has a different skill balance than both Dwarf Rangers and vanilla
Rangers, focused on staying absolutely as far from a fight as possible
with excellent eyes (mild darkvision permanently) and decent Sneaking
skills. In a direct fight, however, they are extraordinarily lightly
armored with only the unusual Hack-Knife machete to protect themselves
with.

This class allows Hollowkin, Humen, and Harpies, since Hollowkin and
Human bandit gangs work together in the lore, and I figured the odd
Harpy would make an EXCELLENT scout if one could ever be hired on.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
More Hollowkin classes are a boon, and demanding players to think about
positioning via a near complete lack of armor feels like an interesting
challenge. A varied skill spread and permanent dark-vision make this
class a good Adventurer pick for those who want to strike an outdoorsy
round out without taking absolutely zero combat skills via a Pilgrim
role.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [X] You tested this on a local server.
- [X] This code did not runtime during testing.
- [X] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->